requirements.txt
edge_tts==6.1.9
fastapi==0.115.14
groq==0.29.0
httpx==0.28.1
python-dotenv==1.1.1
python-socketio==5.13.0
twilio==8.10.0
uvicorn==0.35.0
scikit-learn==1.7.0
langchain==0.3.26
langchain_groq==0.3.5
numpy==2.3.1
